requirements.txt
altgraph==0.17.2
click==8.1.2
Flask==2.1.1
Flask-Login==0.6.0
Flask-SQLAlchemy==2.5.1
gobject==0.1.0
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
packaging==21.3
proxy-tools==0.1.0
pycairo==1.21.0
PyGObject==3.42.1
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.6
pyparsing==3.0.9
PySide2==5.15.2.1
python-dotenv==0.20.0
pywebview==3.6.3
QtPy==2.1.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
shiboken2==5.15.2.1
SQLAlchemy==1.4.35
tk==0.1.0
vext==0.7.6
vext.gi==0.7.4
Werkzeug==2.1.1
zipp==3.8.0
